+++ CHANGELOG.md
@@ -6,12 +6,13 @@ Resolved issues:
 
 - CE-697 Convert datetime with timezone offsets for browser AudioMoth recordings
 
-Other:
+Performance improvements:
 
 - CE-623 Extend amount of time users can stay logged into Arbimon
 - CE-686 Save source file name from Core API into recordings
+- CE-639 Optimize query which gets list if sites on Pattern Matching ROIs list page
 
-## v3.0.24 - May 18, 2021
+## v3.0.24 - May 21, 2021
 
 New features:
 
@@ -29,6 +30,10 @@ Resolved issues:
 - Show spinner while search is processing on Projects page; Hide "Add project" button on Explore projects page. Do not show search form for guests.
 - CE-639 Fix recordings query for project which does not have any sites
 - CE-639 Disable filter buttons on Recordings page while recordings are loading from server
+
+Other:
+
+- CE-686 Save source file name from Core API into recordings
 
 ## v3.0.23 - May 14, 2021